Fix impacts analysis
The impacts analysis relies on draws from the analytical Hessian matrix of the model. The problem is that the Hessian in spatial models is often close to singular, which means it cannot be inverted. spdep gets around this by using a numerical Hessian instead if the inversion fails (this is the warning that you get when you estimate the SDM).

The problem is that then the object that the impacts needs isn't there. So I've tightened the convergence tolerance on the estimation, and everything works. I've also added a helper function to process the impacts statistics in a better way, and that we can adjust as necessary.

This new tolerance could overload your computer if you have a 32-bit processor. If you have one of those, it's time to get a new computer! Or borrow one from AWS, I can show you how.