I went back to double check my tables of log likelihood values to make sure I hadn't entered anything incorrectly. As it turns out, I accidentally printed the physical activity models to htmlreg twice, which explains why the physical activity and mental health likelihood values were the same after adding park access into the models. I fixed the problem and added the correct values to the tables.
I removed the Northeastern census tract from Denver and also removed a handfull of census tracts from Boston that comprised mostly null values. As expected, the log likelihood values changed in Denver and Boston as a result of the smaller number of census tracts, so I updated the tables accordingly.